@@ -130,24 +130,23 @@
     "uuid": "^3.2.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.6.4",
+    "@babel/preset-env": "^7.6.3",
+    "@babel/preset-react": "^7.6.3",
     "ava": "^0.25.0",
     "babel-core": "^6.14.0",
     "babel-jest": "^22.4.3",
-    "babel-loader": "^6.2.0",
-    "babel-plugin-react-transform": "^2.0.0",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-react-css-modules": "^5.2.6",
     "babel-plugin-webpack-alias": "^2.1.1",
     "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-hmre": "^1.0.1",
     "babel-register": "^6.11.6",
     "browser-env": "^3.2.5",
     "color": "^3.0.0",
     "concurrently": "^3.4.0",
     "copy-to-clipboard": "^3.0.6",
     "cross-env": "^5.2.0",
-    "css": "^2.2.4",
-    "css-loader": "^0.19.0",
+    "css-loader": "^1.0.1",
     "devtron": "^1.1.0",
     "dom-storage": "^2.0.2",
     "electron": "3.0.8",
@@ -181,11 +180,12 @@
     "redux-devtools-log-monitor": "^1.4.0",
     "signale": "^1.2.1",
     "standard": "^8.4.0",
-    "style-loader": "^0.12.4",
-    "stylus": "^0.52.4",
-    "stylus-loader": "^2.3.1",
-    "webpack": "^1.12.2",
-    "webpack-dev-server": "^1.12.0"
+    "style-loader": "^1.0.0",
+    "stylus": "0.49.1",
+    "stylus-loader": "^3.0.2",
+    "to-string-loader": "^1.1.5",
+    "webpack": "^4.41.2",
+    "webpack-dev-server": "^3.9.0"
   },
   "optionalDependencies": {
     "grunt-electron-installer-debian": "^0.2.0",
