@@ -1,23 +1,71 @@
 const skeleton = require('./webpack-skeleton')
 const path = require('path')
 
+// stylus: {
+//   use: [require('nib')()],
+// import: [
+//     '~nib/lib/nib/index.styl',
+//     path.join(__dirname, 'browser/styles/index.styl')
+//   ]
+// },
+
 var config = Object.assign({}, skeleton, {
   module: {
-    loaders: [
-      {
-        test: /\.js?$/,
-        exclude: /node_modules/,
-        loader: 'babel?cacheDirectory'
-      },
+    rules: [
       {
         test: /\.styl$/,
-        exclude: /(node_modules|bower_components)/,
-        loader: 'style!css?modules&importLoaders=1&localIdentName=[name]__[local]___[path]!stylus?sourceMap'
+        use: [
+          {
+            loader: 'style-loader' // creates style nodes from JS strings
+          },
+          {
+            loader: 'css-loader?modules&importLoaders=1&localIdentName=[path]___[name]__[local]___[hash:base64:5]' // translates CSS into CommonJS
+          },
+          {
+            loader: 'stylus-loader', // compiles Stylus to CSS
+            options: {
+              use: [require('nib')()],
+              import: ['~nib/lib/nib/index.styl',
+                path.join(__dirname, 'browser/styles/index.styl')
+              ]
+            }
+          }
+        ]
       },
+      // {
+      //   test: /\.styl$/,
+      //   exclude: /(node_modules|bower_components)/,
+      //   use: [
+      //     {
+      //       loader: 'stylus-loader',
+      //       options: {
+      //         use: [require('nib')()],
+      //         import: ['~nib/lib/nib/index.styl',
+      //           path.join(__dirname, 'browser/styles/index.styl')
+      //         ]
+      //       }
+      //     }
+      //   ]
+      // },
       {
-        test: /\.json$/,
-        loader: 'json'
+        test: /\.m?js$/,
+        exclude: /(node_modules|bower_components)/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-env']
+          }
+        }
       }
+      // {
+      //   test: /\.css$/i,
+      //   use: ['style-loader', 'css-loader']
+      // },
+      // {
+      //   test: /\.styl$/,
+      //   exclude: /(node_modules|bower_components)/,
+      //   use: 'css-loader?modules&importLoaders=1&localIdentName=[path]___[name]__[local]___[hash:base64:5]'
+      // },
     ]
   },
   output: {
@@ -27,7 +75,6 @@ var config = Object.assign({}, skeleton, {
     libraryTarget: 'commonjs2',
     publicPath: 'http://localhost:8080/assets/'
   },
-  debug: true,
   devtool: 'cheap-module-eval-source-map',
   devServer: {
     port: 8080,
